+++ openquake/calculators/tests/event_based_risk_test.py
@@ -554,7 +554,7 @@ agg_id
         # the parent has aggregate_by = NAME_1, NAME_2, taxonomy
         oq = parent['oqparam']
         oq.__dict__['aggregate_by'] = ['NAME_1']
-        job_id = logs.init('nojob', logging.INFO)  # requires the DbServer
+        job_id = logs.init('calc', logging.INFO)  # requires the DbServer
         prc = PostRiskCalculator(oq, job_id)
         oq.hazard_calculation_id = parent.calc_id
         with mock.patch.dict(os.environ, {'OQ_DISTRIBUTE': 'no'}):
